+++ doc/changelog.rst
@@ -7,6 +7,9 @@ Changelog
 New
 ~~~
 
+- Expose additional integer and remainder-related functions
+  for :cpp:class:`~mppp::real`
+  (`#254 <https://github.com/bluescarni/mppp/pull/254>`__).
 - Add a tutorial for :cpp:class:`~mppp::real`
   (`#252 <https://github.com/bluescarni/mppp/pull/252>`__).
 - Expose all the constants provided by MPFR in the